fixed import error and transfered env function to config
