+++ API/changesets/routers.py
@@ -2,7 +2,7 @@ import psycopg2
 
 from psycopg2 import sql
 from psycopg2.extras import DictCursor
-from src.galaxy import get_db_connection_params
+from src.galaxy.config import get_db_connection_params
 from . import ChangesetResult, FilterParams
 from .utils import geom_filter_subquery
 from fastapi import APIRouter
