+++ src/galaxy/app.py
@@ -34,7 +34,7 @@ from json import loads as json_loads
 from geojson import Feature, FeatureCollection, Point
 from io import StringIO
 
-from .config import config
+from .config import get_db_connection_params
 
 def print_psycopg2_exception(err):
     """ 
@@ -139,7 +139,7 @@ class Underpass:
     """This class connects to underpass database and responsible for all the underpass related functionality"""
 
     def __init__(self, parameters=None):
-        self.database = get_db_connection_params("UNDERPASS")
+        self.database = Database(get_db_connection_params("UNDERPASS"))
         # self.database = Database(dict(config.items("UNDERPASS")))
         self.con, self.cur = self.database.connect()
         self.params = parameters
@@ -178,7 +178,7 @@ class Insight:
     """This class connects to Insight database and responsible for all the Insight related functionality"""
 
     def __init__(self, parameters=None):
-        self.database = get_db_connection_params("INSIGHTS")
+        self.database = Database(get_db_connection_params("INSIGHTS"))
         # self.database = Database(dict(config.items("INSIGHTS_PG")))
         self.con, self.cur = self.database.connect()
         self.params = parameters
@@ -340,7 +340,7 @@ class Output:
 
 class UserStats:
     def __init__(self):
-        self.db = get_db_connection_params("INSIGHTS")
+        self.db = Database(get_db_connection_params("INSIGHTS"))
         # self.db = Database(dict(config.items("INSIGHTS_PG")))
         self.con, self.cur = self.db.connect()
 
@@ -382,7 +382,7 @@ class UserStats:
 
 class DataQualityHashtags:
     def __init__(self, params: DataQualityHashtagParams):
-        self.db = get_db_connection_params("UNDERPASS")
+        self.db = Database(get_db_connection_params("UNDERPASS"))
         # self.db = Database(dict(config.items("UNDERPASS")))
         self.con, self.cur = self.db.connect()
         self.params = params
@@ -451,7 +451,7 @@ class DataQuality:
     """
 
     def __init__(self, parameters, inputtype):
-        self.db = get_db_connection_params("UNDERPASS")
+        self.db = Database(get_db_connection_params("UNDERPASS"))
         # self.db = Database(dict(config.items("UNDERPASS")))
         self.con, self.cur = self.db.connect()
         self.inputtype = inputtype
@@ -527,7 +527,7 @@ class OrganizationHashtags :
     """[Class responsible for Organization Hashtag data API]
     """
     def __init__(self, params: OrganizationHashtagParams):
-        self.db = get_db_connection_params("INSIGHTS")
+        self.db = Database(get_db_connection_params("INSIGHTS"))
         # self.db = Database(dict(config.items("INSIGHTS_PG")))
         self.con, self.cur = self.db.connect()
         self.params = params
